// ui/src/utils/excel/interface.ts
export interface IExcelExportCourses {
    courseName: string;
    branch: string;
    networks: string;
    courseType: string;
    courseACAId: string;
    baseName: string;
    year: number;
    startDate: string;
    endDate: string;
    bedrooms: string;
    offices: string;
    classes: string;
    unit: string;
    courseSAPId: string;
    profession: string;
    RAKAZCourseDuration: number;
    actualCourseDuration: number;
    basicTrainingStartDate: string;
    basicTrainingEndDate: string;
    receivanceDate: string;
    enlistmentDate?: string;
    iturMale: number;
    iturFemale: number;
    specialMale: number;
    specialFemale: number;
    totalMale: number;
    totalFemale: number;
}